backend communicates with frontend

+++ backend/src/main/java/com/carproject/repo/BrandRepository.java
@@ -3,5 +3,5 @@ package com.carproject.repo;
 import com.carproject.domain.CarBrand;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BrandRepository extends JpaRepository<CarBrand, Long> {
+public interface BrandRepository extends JpaRepository<CarBrand, Integer> {
 }
